move fetch types to web-globals
