+++ types/node/index.d.ts
@@ -38,6 +38,7 @@
 
 // Definitions for Node.js modules that are not specific to any version of TypeScript:
 /// <reference path="globals.d.ts" />
+/// <reference path="web-globals/fetch.d.ts" />
 /// <reference path="web-globals/navigator.d.ts" />
 /// <reference path="assert.d.ts" />
 /// <reference path="assert/strict.d.ts" />
